+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+ - Bug where missing URL / method fields caused a crash
+ - Issue where only the first validation error is printed
 
 ## [0.4.4] - 2019-02-17
 ### Fixed